Revert "Migrating AnimalRequirements-specific unit tests to dedicated test file."
